third_party/blink/web_tests/http/tests/activedomobject/media-expected.txt

Tests that reparenting media elements also reparents ActiveDOMObject.

Before Reparenting
PASS: internals.pausableObjectCount(document) should be '0' and is.
PASS: internals.pausableObjectCount(iframe) should be '1' and is.
After Reparenting
PASS: internals.pausableObjectCount(document) should be '1' and is.
PASS: internals.pausableObjectCount(iframe) should be '0' and is.
